refactor oaoo tests
